+++ src/view/group/problem-list/RegisterForm/index.css.ts
