+++ src/view/group/problem-list/RegisterForm/schema.ts
